STEP 2 - 내 번호와 로또 번호 맞춰보기 (파일 나누기)
